@@ -21,13 +21,15 @@
     "classnames": "^2.2.5"
   },
   "peerDependencies": {
-    "react": "^15.0.0"
+    "react": ">=15.0.0"
   },
   "devDependencies": {
     "babel-core": "^6.9.0",
     "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-react": "^6.5.0",
+    "create-react-class": "^15.6.3",
     "css-loader": "~0.23.1",
     "es5-shim": "~4.5.8",
     "eslint": "^2.10.2",
@@ -42,9 +44,9 @@
     "karma-webpack": "~1.7.0",
     "node-sass": "^3.7.0",
     "phantomjs-prebuilt": "^2.1.7",
-    "react": "^15.1.0",
-    "react-addons-test-utils": "^15.1.0",
-    "react-dom": "^15.1.0",
+    "prop-types": "^15.6.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "sass-loader": "~3.2.0",
     "style-loader": "~0.13.1",
     "webpack": "^1.13.1",
